Further extractions and simplifications.
